requirements.txt
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
dill==0.2.7.1
dnspython==1.12.0
enum34==1.1.6
future==0.16.0
futures==3.1.1
google-api-core==0.1.1
google-auth==1.2.1
google-cloud-core==0.28.0
google-cloud-speech==0.30.0
google-gax==0.15.16
googleapis-common-protos==1.5.3
grpcio==1.7.0
idna==2.6
ms-cognitive-speaker-recognition==0.1
numpy==1.13.3
ply==3.8
protobuf==3.5.0.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pycrypto==2.6.1
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
